code clean up
