+++ .gitignore
@@ -8,3 +8,4 @@ build/
 openssl.exe
 libcrypto-3-x64.dll
 libssl-3-x64.dll
+/.vs
