--- websmtp/Endpoints/ChangePasswordViewModel.cs
@@ -1,13 +1,0 @@
-namespace websmtp;
-
-public static partial class MessagesEndpoints
-{
-    public class ChangePasswordViewModel
-    {
-        public int UserId { get; set; }
-        public string CurrentPassword { get; set; } = string.Empty;
-        public string NewPassword { get; set; } = string.Empty;
-        public string ConfirmPassword { get; set; } = string.Empty;
-    }
-
-}
