+++ websmtp/Endpoints/Models/AddMailboxRequest.cs
@@ -1,0 +1,8 @@
+namespace websmtp;
+
+public class AddMailboxRequest
+{
+    public int UserId { get; set; }
+    public string DisplayName { get; set; } = string.Empty;
+    public string Email { get; set; } = string.Empty;
+}
