+++ websmtp/Endpoints/Models/AddUserRequest.cs
@@ -1,0 +1,9 @@
+﻿namespace websmtp;
+
+public class AddUserRequest
+{
+    public string Username { get; set; } = string.Empty;
+    public string RealName { get; set; } = string.Empty;
+    public string Email { get; set; } = string.Empty;
+    public string Password { get; set; } = string.Empty;
+}
