+++ websmtp/Endpoints/Models/ChangePasswordViewModel.cs
@@ -1,0 +1,9 @@
+namespace websmtp;
+
+public class ChangePasswordViewModel
+{
+    public int UserId { get; set; }
+    public string CurrentPassword { get; set; } = string.Empty;
+    public string NewPassword { get; set; } = string.Empty;
+    public string ConfirmPassword { get; set; } = string.Empty;
+}
