+++ websmtp/Endpoints/Models/ChangeUserPasswordRequest.cs
@@ -1,0 +1,8 @@
+﻿namespace websmtp;
+
+public class ChangeUserPasswordRequest
+{
+    public int UserId { get; set; }
+    public string NewPassword { get; set; } = string.Empty;
+    public string ConfirmPassword { get; set; } = string.Empty;
+}
