+++ websmtp/Endpoints/Models/ChangeUsernameRequest.cs
@@ -1,0 +1,7 @@
+﻿namespace websmtp;
+
+public class ChangeUsernameRequest
+{
+    public int UserId { get; set; }
+    public string NewUsername { get; set; } = string.Empty;
+}
