+++ websmtp/Endpoints/Models/OtpValidateViewModel.cs
@@ -1,0 +1,7 @@
+namespace websmtp;
+
+
+public class OtpValidateViewModel
+{
+    public string Otp { get; set; } = string.Empty;
+}
