+++ websmtp/Endpoints/Models/RemoveUserMailboxRequest.cs
@@ -1,0 +1,7 @@
+﻿namespace websmtp;
+
+public class RemoveUserMailboxRequest
+{
+    public int UserId { get; set; }
+    public int MailboxId { get; set; }
+}
