+++ websmtp/Endpoints/Models/TrainRequest.cs
@@ -1,0 +1,7 @@
+﻿namespace websmtp;
+
+public class TrainRequest
+{
+    public List<Guid> MsgsIds { get; set; } = [];
+    public bool Spam { get; set; }
+}
