--- websmtp/Endpoints/OtpValidateViewModel.cs
@@ -1,10 +1,0 @@
-namespace websmtp;
-
-public static partial class MessagesEndpoints
-{
-    public class OtpValidateViewModel
-    {
-        public string Otp { get; set; } = string.Empty;
-    }
-
-}
