+++ websmtp/Pages/Index.cshtml.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
-using websmtp.services;
+using websmtp.Services.Models;
 
 namespace websmtp.Pages;
 
@@ -19,7 +19,7 @@ public class IndexModel : PageModel
     public IndexModel(ILogger<IndexModel> logger,
     IReadableMessageStore messageStore)
     {
-        _logger = logger ?? throw new ArgumentNullException(nameof(_logger));
+        _logger = logger ?? throw new ArgumentNullException(nameof(logger));
         _messageStore = messageStore ?? throw new ArgumentNullException(nameof(messageStore));
     }
 
@@ -27,33 +27,15 @@ public class IndexModel : PageModel
         [FromQuery] int page = 1,
         [FromQuery] int perPage = 1000)
     {
-        switch (Mailbox)
+        Listing = Mailbox switch
         {
-            default:
-            case "inbox":
-                Listing = _messageStore.Latest(page, perPage, true, false, false, false, false, Filter);
-                break;
-
-            case "all":
-                Listing = _messageStore.Latest(page, perPage, false, false, false, false, false, Filter);
-                break;
-
-            case "favorites":
-                Listing = _messageStore.Latest(page, perPage, false, false, false, true, false, Filter);
-                break;
-
-            case "spam":
-                Listing = _messageStore.Latest(page, perPage, false, false, true, false, false, Filter);
-                break;
-
-            case "sent":
-                Listing = _messageStore.Latest(page, perPage, false, false, false, false, true, Filter);
-                break;
-
-            case "trash":
-                Listing = _messageStore.Latest(page, perPage, false, true, true, false, false, Filter);
-                break;
-        }
+            "all" => _messageStore.Latest(page, perPage, false, false, false, false, false, Filter),
+            "favorites" => _messageStore.Latest(page, perPage, false, false, false, true, false, Filter),
+            "spam" => _messageStore.Latest(page, perPage, false, false, true, false, false, Filter),
+            "sent" => _messageStore.Latest(page, perPage, false, false, false, false, true, Filter),
+            "trash" => _messageStore.Latest(page, perPage, false, true, true, false, false, Filter),
+            _ => _messageStore.Latest(page, perPage, true, false, false, false, false, Filter),
+        };
         return Page();
     }
 }
