+++ websmtp/Pages/Shared/Sidebar.cshtml
@@ -1,4 +1,4 @@
-﻿@model websmtp.services.ListResult
+﻿@model websmtp.Services.Models.ListResult
 <nav class="sidebar border-right d-none d-xl-block">
     <a id="logo--link" href="/inbox"
         class="h2 py-2 d-flex align-items-center justify-content-center text-decoration-none">
