+++ websmtp/Pages/View.cshtml.cs
@@ -10,16 +10,13 @@ public class MessageViewModel : PageModel
 {
     [FromQuery] public Guid MsgId { get; set; } = Guid.Empty;
     [FromQuery] public bool ShowRaw { get; set; } = false;
-    private readonly ILogger<IndexModel> _logger;   
     private readonly IReadableMessageStore _messageStore;
 
     public Message Message { get; set; }
 
     public MessageViewModel(
-        ILogger<IndexModel> logger,
         IReadableMessageStore messageStore)
     {
-        _logger = logger ?? throw new ArgumentNullException(nameof(_logger));
         _messageStore = messageStore ?? throw new ArgumentNullException(nameof(messageStore));
         Message = new Message();
     }
